# apps/api/src/routes/v1/onboarding/index.ts
import type { FastifyInstance } from 'fastify';
import { randomUUID } from 'node:crypto';

const sessions = new Map<string, any>();
const now = () => new Date().toISOString();
const nextAction = (s:any) => s.status === 'completed'
  ? { type:'done', step:'completed', label:'Onboarding complete', required_fields:[] }
  : { type: s.status === 'pending_review' ? 'wait' : 'collect_input', step: s.current_step ?? 'organization', label:`Complete ${s.current_step ?? 'organization'}`, required_fields:[], ...(s.status === 'pending_review' ? { retry_after_seconds:30 } : {}) };
const publicSession = (s:any) => ({ id:s.id, object:'onboarding_session', livemode:s.livemode, tenant_id:s.tenant_id, environment_id:s.environment_id, status:s.status, current_step:s.current_step, step_state:s.step_state, kyb_decision_id:s.kyb_decision_id, evidence_file_ids:s.evidence_file_ids, first_api_key_id:s.first_api_key_id, first_transfer_intent_id:s.first_transfer_intent_id, completion_event_id:s.completion_event_id, created:s.created_at, updated:s.updated_at, completed_at:s.completed_at, next_action:nextAction(s) });
const requireScope = (r:any, scope:string) => { const scopes = r.auth?.scopes ?? ['read:onboarding','write:onboarding','*']; if (!scopes.includes('*') && !scopes.includes(scope)) { const e:any = new Error('missing onboarding scope'); e.statusCode = 403; throw e; } };
export async function onboardingRoutes(s: FastifyInstance) {
  s.post('/onboarding', async r => { requireScope(r,'write:onboarding'); const body:any=r.body??{}; const key=(r.headers['idempotency-key'] as string|undefined) ?? ''; const existing=[...sessions.values()].find(x=>x.idempotency_key===key && key); if(existing) return publicSession(existing); const ts=now(); const rec={ id:`onb_${randomUUID().replaceAll('-','')}`, idempotency_key:key, livemode:r.auth?.livemode??false, tenant_id:r.accountId??'tnt_demo', environment_id:body.environment_id??'env_test', status:'in_progress', current_step:'organization', step_state: body.organization ? { organization: body.organization } : {}, evidence_file_ids:[], created_at:ts, updated_at:ts }; sessions.set(rec.id, rec); return publicSession(rec); });
  s.get('/onboarding/:id', async r => { requireScope(r,'read:onboarding'); const rec=sessions.get((r.params as any).id); if(!rec) return s.httpErrors.notFound('onboarding session not found'); return publicSession(rec); });
  s.post('/onboarding/:id/advance', async r => { requireScope(r,'write:onboarding'); const rec=sessions.get((r.params as any).id); if(!rec) return s.httpErrors.notFound('onboarding session not found'); const body:any=r.body??{}; rec.step_state={...rec.step_state,[body.step]:body.input??{}}; const order=['organization','kyb','sandbox','api_key','webhook','first_transfer','completed']; if(body.step==='kyb' && body.input?.outage){ rec.status='pending_review'; rec.current_step='kyb'; } else if(body.step==='first_transfer' && body.input?.projected){ rec.status='completed'; rec.current_step='completed'; rec.first_transfer_intent_id=body.input.transfer_intent_id??rec.first_transfer_intent_id; rec.completion_event_id=body.input.event_id??`evt_${Date.now()}`; rec.completed_at=now(); } else { const i=order.indexOf(body.step); rec.status='in_progress'; rec.current_step=order[Math.min(i+1, order.length-1)]; } rec.updated_at=now(); return publicSession(rec); });
  s.post('/onboarding/:id/cancel', async r => { requireScope(r,'write:onboarding'); const rec=sessions.get((r.params as any).id); if(!rec) return s.httpErrors.notFound('onboarding session not found'); rec.status='canceled'; rec.updated_at=now(); return publicSession(rec); });
  s.get('/onboarding/:id/next_action', async r => { requireScope(r,'read:onboarding'); const rec=sessions.get((r.params as any).id); if(!rec) return s.httpErrors.notFound('onboarding session not found'); return nextAction(rec); });
}