# packages/api-contracts/src/codegen/schemas.gen.ts
import { z } from 'zod';
import { AccountSchema, AccountCreateRequest, AccountUpdateRequest, AccountListResponse, AssetSchema, AssetCreateRequest, AssetUpdateRequest, AssetListResponse, BalanceSchema, BalanceListResponse, HoldingSchema, HoldingListResponse, IssueIntentSchema, IssueIntentCreateRequest, IssueIntentListResponse, RedeemIntentSchema, RedeemIntentCreateRequest, RedeemIntentListResponse, TransferIntentSchema, TransferIntentCreateRequest, TransferIntentListResponse, HoldSchema, HoldCreateRequest, HoldReleaseRequest, HoldListResponse, OperationSchema, OperationListResponse, EventSchema, EventListResponse, WebhookEndpointSchema, WebhookEndpointCreateRequest, WebhookEndpointUpdateRequest, RotateSecretResponse, WebhookEndpointListResponse, ApiKeySchema, ApiKeyCreateRequest, ApiKeyCreateResponse, ApiKeyListResponse, OnboardingSession, OnboardingSessionCreateRequest, OnboardingSessionAdvanceRequest, OnboardingNextAction, SearchQuery, SearchResponse, ExportJob, ExportCreateRequest, ExportDownloadUrlResponse, ExportJobListResponse, ReportTemplate, ReportTemplateCreateRequest, ReportTemplateListResponse, ReportTemplateRunResponse, ExportDestination, ExportDestinationCreateRequest, ExportDestinationListResponse } from '../schemas/index.js';

export type Account = z.infer<typeof AccountSchema>;
export type AccountCreate = z.infer<typeof AccountCreateRequest>;
export type AccountUpdate = z.infer<typeof AccountUpdateRequest>;
export type AccountList = z.infer<typeof AccountListResponse>;
export type Asset = z.infer<typeof AssetSchema>;
export type AssetCreate = z.infer<typeof AssetCreateRequest>;
export type AssetUpdate = z.infer<typeof AssetUpdateRequest>;
export type AssetList = z.infer<typeof AssetListResponse>;
export type Balance = z.infer<typeof BalanceSchema>;
export type BalanceList = z.infer<typeof BalanceListResponse>;
export type Holding = z.infer<typeof HoldingSchema>;
export type HoldingList = z.infer<typeof HoldingListResponse>;
export type IssueIntent = z.infer<typeof IssueIntentSchema>;
export type IssueIntentCreate = z.infer<typeof IssueIntentCreateRequest>;
export type IssueIntentList = z.infer<typeof IssueIntentListResponse>;
export type RedeemIntent = z.infer<typeof RedeemIntentSchema>;
export type RedeemIntentCreate = z.infer<typeof RedeemIntentCreateRequest>;
export type RedeemIntentList = z.infer<typeof RedeemIntentListResponse>;
export type TransferIntent = z.infer<typeof TransferIntentSchema>;
export type TransferIntentCreate = z.infer<typeof TransferIntentCreateRequest>;
export type TransferIntentList = z.infer<typeof TransferIntentListResponse>;
export type Hold = z.infer<typeof HoldSchema>;
export type HoldCreate = z.infer<typeof HoldCreateRequest>;
export type HoldRelease = z.infer<typeof HoldReleaseRequest>;
export type HoldList = z.infer<typeof HoldListResponse>;
export type Operation = z.infer<typeof OperationSchema>;
export type OperationList = z.infer<typeof OperationListResponse>;
export type Event = z.infer<typeof EventSchema>;
export type EventList = z.infer<typeof EventListResponse>;
export type WebhookEndpoint = z.infer<typeof WebhookEndpointSchema>;
export type WebhookEndpointCreate = z.infer<typeof WebhookEndpointCreateRequest>;
export type WebhookEndpointUpdate = z.infer<typeof WebhookEndpointUpdateRequest>;
export type RotateSecret = z.infer<typeof RotateSecretResponse>;
export type WebhookEndpointList = z.infer<typeof WebhookEndpointListResponse>;
export type ApiKey = z.infer<typeof ApiKeySchema>;
export type ApiKeyCreate = z.infer<typeof ApiKeyCreateRequest>;
export type ApiKeyCreate2 = z.infer<typeof ApiKeyCreateResponse>;
export type ApiKeyList = z.infer<typeof ApiKeyListResponse>;
export type OnboardingSession = z.infer<typeof OnboardingSession>;
export type OnboardingSessionCreate = z.infer<typeof OnboardingSessionCreateRequest>;
export type OnboardingSessionAdvance = z.infer<typeof OnboardingSessionAdvanceRequest>;
export type OnboardingNextAction = z.infer<typeof OnboardingNextAction>;
export type SearchQuery = z.infer<typeof SearchQuery>;
export type Search = z.infer<typeof SearchResponse>;
export type ExportJob = z.infer<typeof ExportJob>;
export type ExportCreate = z.infer<typeof ExportCreateRequest>;
export type ExportDownloadUrl = z.infer<typeof ExportDownloadUrlResponse>;
export type ExportJobList = z.infer<typeof ExportJobListResponse>;
export type ReportTemplate = z.infer<typeof ReportTemplate>;
export type ReportTemplateCreate = z.infer<typeof ReportTemplateCreateRequest>;
export type ReportTemplateList = z.infer<typeof ReportTemplateListResponse>;
export type ReportTemplateRun = z.infer<typeof ReportTemplateRunResponse>;
export type ExportDestination = z.infer<typeof ExportDestination>;
export type ExportDestinationCreate = z.infer<typeof ExportDestinationCreateRequest>;
export type ExportDestinationList = z.infer<typeof ExportDestinationListResponse>;